# Entity struct as record and skipinitialization