--- a/Arch/Core/Utils/NetStandardHack.cs
+++ b/Arch/Core/Utils/NetStandardHack.cs
@@ -1,0 +1,6 @@
+﻿#if NETSTANDARD2_1_OR_GREATER
+namespace System.Runtime.CompilerServices
+{
+    internal static class IsExternalInit { }
+}
+#endif