@genaray 2022-12-20
Thanks ! :) 

I would actually prefer to keep the `Entity` as it is in the current version. The default c# hash method is not the fastest ( as far as i know ) and the current implementation is the fastest possible one ( atleast i saw it multiple time advised on several different [sources ](https://stackoverflow.com/questions/263400/what-is-the-best-algorithm-for-overriding-gethashcode)). Furthermore it does not cause any adress jumping, relying on records and default hashes often cause adress jumping.

However, your `[SkipInitLocals]` approach is quite interesting, that could be merged. Probably there more places where it makes sense, for example inside queries, add, remove or other method groups ^^